L1Algo init interface updates

+++ reco/L1/L1Algo/L1Def.h
@@ -65,10 +65,7 @@ typedef int index_type;
 
 /// Hash for unordered_map with enum class keys
 struct EnumClassHash {
-  template <typename T> size_t operator()(T t) const 
-  {
-    return static_cast<size_t>(t);
-  }
+  template <typename T> int operator()(T t) const { return static_cast<int>(t); }
 };
 
 #endif
